requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
asttokens==2.4.1
async-property==0.2.2
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
dj-svg==0.3.1
Django==5.1.5
django-allauth==65.3.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-multiselectfield==0.1.13
django-permissionedforms==0.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rosetta==0.10.1
django-taggit==6.1.0
django-treebeard==4.7.1
django-unfold==0.40.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
draftjs-exporter==5.0.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
executing==2.1.0
filelock==3.13.4
filetype==1.2.0
Flask==3.0.2
frozenlist==1.5.0
h11==0.14.0
h3==3.7.7
httpcore==1.0.7
httpx==0.28.1
idna==3.7
inflection==0.5.1
ipdb==0.13.13
ipython==8.31.0
isodate==0.7.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jwcrypto==1.5.6
l18n==2021.3
laces==0.1.2
ldap3==2.9.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msal==1.32.0
multidict==6.1.0
mysqlclient==2.2.7
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
phonenumbers==8.13.47
pillow==11.1.0
pillow_heif==0.21.0
platformdirs==4.2.1
polib==1.2.0
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pycryptodomex==3.11.0
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-keycloak==5.2.0
python-weather==2.0.7
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==79.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
telepath==0.3.1
termcolor==2.5.0
timezonefinder==6.2.0
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.0
wagtail==6.3.2
wagtail-localize==1.11.2
wagtail-modeladmin==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
Willow==1.9.0
XlsxWriter==3.1.2
yarl==1.18.3
zeep==4.2.1
